DHSCFT-422: Correct imports after separating files
